refactor(Equipment): Extracted equipment func --
Extracted equipment functionality from entity to equipmentmanager(classname TBD).

I'm trying to break out some parts from the entity class due to the sheer size of it.
