+++ src/GameObjects/Entities/Entity.java
@@ -1,6 +1,7 @@
 package GameObjects.Entities;
 
 import GameObjects.Items.Equipment;
+import GameObjects.Items.EquipmentManager;
 import GameObjects.Items.EquipmentSlot;
 import GameObjects.Items.Inventory;
 import Global.Utility;
@@ -12,7 +13,7 @@ public class Entity {
     private static final int DEFAULT_STAT = 3;
 
     // TODO break out the equipment slot/hashmap of equipment into its own class. TT
-    private final HashMap<EquipmentSlot, Equipment> equipment;
+    private final EquipmentManager equipmentManager;
     private final Inventory inventory;
     private final String name;
     protected int health;
@@ -36,10 +37,7 @@ public class Entity {
 
         // Creates an eq "space" for each type and assigns null to them.
         // so if one entity is created they will have "nothing" on them so to speak.
-        this.equipment = new HashMap<>();
-        for (EquipmentSlot slot : EquipmentSlot.values()) {
-            equipment.put(slot, null);
-        }
+        this.equipmentManager = new EquipmentManager();
     }
 
     // Constructor with base stats (Standard NPC)
@@ -84,10 +82,8 @@ public class Entity {
         return level;
     }
 
-    public Equipment getEquipment(EquipmentSlot slot) {return equipment.get(slot);}
-
-    public HashMap<EquipmentSlot, Equipment> getEquipmentList() {
-        return equipment;
+    public EquipmentManager getEquipmentList() {
+        return equipmentManager;
     }
 
     public boolean isDead() {
@@ -170,28 +166,6 @@ public class Entity {
         } else {
             System.out.println(name + " took " + damageValue + " damage. Current health: " + health);
         }
-    }
-
-    // method to call to add and swap equipments from X slot.
-    public Equipment equipItem(Equipment eq) {
-        EquipmentSlot eqSlot = eq.getEQSlot();
-        Equipment previousEQ = equipment.get(eqSlot);
-        equipment.put(eqSlot, eq);
-
-        if (previousEQ != null) {
-            System.out.println("You put " + previousEQ.getName() + " in your inventory.");
-        }
-        return previousEQ;
-    }
-
-    // method to call to remove eq from X slot
-    public Equipment unequipItem(EquipmentSlot slot) {
-        if (equipment.containsKey(slot)) {
-            Equipment removedItem = equipment.get(slot);
-            equipment.put(slot, null);
-            return removedItem;
-        }
-        return null;
     }
 
 }
