+++ src/GameObjects/Entities/PlayerCharacter.java
@@ -18,7 +18,6 @@ public class PlayerCharacter extends Entity {
 
     public PlayerCharacter(String name, int health, int str, int dex, int intel) {
         super(name, health, 1, str, dex, intel);
-
         currentZone = ZoneManager.getZone(ZoneType.TAVERN);
     }
