+++ src/GameObjects/Items/Equipment.java
@@ -30,7 +30,7 @@ public class Equipment extends Item {
         System.out.print("Do you want to equip this item? (Y/N): ");
         if (Utility.checkYesOrNo(sc)) {
             System.out.println("You equipped " + getName());
-            player.getInventory().addItem(player.equipItem(this));
+            player.getInventory().addItem(player.getEquipmentList().equipItem(this));
             player.getInventory().removeItem(this);
         } else {
             System.out.println("You decided not to equip the item.");
